remove ClientWrapper
